+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changelog
 2.3.0 (unreleased)
 ------------------
 
+- #2147 Remove stale function workflow.getReviewHistory
 - #2146 Fix "No object found for UID: <laboratory_uid>" in report preview
 - #2145 Crop page navigation for DX reference widget
 - #2143 Fix Traceback when using readonly decorator for objects w/o __name__